requirements.txt
django
django-filter
djangorestframework

django-allauth==0.55.0
dj-rest-auth
python-dateutil
xhtml2pdf
django-phonenumber-field

markdown
crispy-bootstrap4
django-filter
django-resized
python-decouple
django_notifications_hq
django-crispy-forms
pillow
six
Faker

requests
gunicorn
psycopg2-binary

django-phonenumber-field
django-phonenumbers
stripe
django-browser-reload
